+++ examples/editor/src/PluginsConfig.js
@@ -77,12 +77,17 @@ export default {
       console.log('consumer wants to upload custom video');
       const customVideoUrl = 'http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4';
       setTimeout(() => {
+        // You can provide either URL or pathname. If pathname is provided getVideoUrl will be invoked to form full URL.
         updateEntity({ data: { url: customVideoUrl } });
         //updateEntity({ error: { msg: 'Upload Failed' } });
         console.log('consumer uploaded ', customVideoUrl);
       }, 500);
     },
     enableCustomUploadOnMobile: true,
+    // Function is invoked when rendering video which only has pathname. You should take the pathname and form a full URL.
+    // getVideoUrl: src => {
+    //   return `https://video.wixstatic.com${src.pathname}`;
+    // },
   },
   [GIPHY_TYPE]: {
     giphySdkApiKey: "JfQziSTdnSwDZcD3w8DpUL4LMFu3zBgU"